David's Photoshop/requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
conda==4.5.9
cryptography==2.2.2
cycler==0.10.0
idna==2.6
imageio==2.3.0
kiwisolver==1.0.1
matplotlib==2.2.2
mkl-fft==1.0.4
mkl-random==1.0.1
numpy==1.15.0
olefile==0.45.1
Pillow==5.2.0
pycosat==0.6.3
pycparser==2.18
PyMaxflow==1.2.9
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
requests==2.18.4
ruamel-yaml==0.15.37
six==1.11.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1
urllib3==1.22
